+++ changelog.md
@@ -1,3 +1,9 @@
+## Next
+
+### Bug fixes
+
+- [jss-plugin-expand] Fix attributes spread for `border-bottom`, `border-top`, `border-left` and `border-right`([#1083](https://github.com/cssinjs/jss/pull/1083))
+
 ## 10.0.0-alpha.16 (2019-3-24)
 
 ### Bug fixes